完成眩晕特效, 优化命名

+++ SERVER/GameServer/PlayerSystem/Player.cs
@@ -110,8 +110,8 @@ namespace GameServer.PlayerSystem
             Level = newLevel;
             EntityAttributeEntrySync(EntityAttributeEntryType.Level, newLevel);
             AttributeManager.Recalculate();
-            EntityAttributeEntrySync(EntityAttributeEntryType.MaxHp, AttributeManager.Final.MaxHp);
-            EntityAttributeEntrySync(EntityAttributeEntryType.MaxMp, AttributeManager.Final.MaxMp);
+            EntityAttributeEntrySync(EntityAttributeEntryType.MaxHp, (int)AttributeManager.Final.MaxHp);
+            EntityAttributeEntrySync(EntityAttributeEntryType.MaxMp, (int)AttributeManager.Final.MaxMp);
             EntityAttributeEntrySync(EntityAttributeEntryType.MaxExp, CalculateExp(Level));
             ChangeHp(AttributeManager.Final.MaxHp);
             ChangeMp(AttributeManager.Final.MaxMp);
